frontend/requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

streamlit
python-dotenv
requests
google-cloud-storage
pandas